Modificaas varias clases agregado el elemnto venta, comenzado el diseño de controladores

+++ src/negocio/entidades/Producto.java
@@ -10,21 +10,21 @@ public abstract class Producto{
     private String      NOMBRE_;
     
     private String      descripcion_;
-    private float       precioVenta_;
-    private float       precioCompra_;
+    private double       precioVenta_;
+    private double       precioCompra_;
 
-    public Producto(int idProducto, String nombre, float precio, float costo) {
+    public Producto(int idProducto, String nombre, double precio, double costo) {
         this.ID_PRODUCTO_   = idProducto;
         this.NOMBRE_        = nombre;
         this.precioVenta_   = precio;
         this.precioCompra_  = costo;
     }
 
-    public float obtenerCosto() {
+    public double obtenerCosto() {
         return precioCompra_;
     }
 
-    public void establecerCosto(float costo) {
+    public void establecerCosto(double costo) {
         this.precioCompra_ = costo;
     }
 
@@ -36,11 +36,11 @@ public abstract class Producto{
         this.descripcion_ = descripcion;
     }
 
-    public float obtenerPrecio() {
+    public double obtenerPrecio() {
         return precioVenta_;
     }
 
-    public void establecerPrecio(float precio) {
+    public void establecerPrecio(double precio) {
         this.precioVenta_ = precio;
     }
